+++ types/graph/utml/parseresult-utml.go
@@ -114,8 +114,8 @@ type UTMLFieldOrMethod struct {
 }
 
 type UTMLXY struct {
-	X int32 `json:"x"`
-	Y int32 `json:"y"`
+	X float64 `json:"x"`
+	Y float64 `json:"y"`
 }
 
 type UTMLEdgeXYOrOffsetPosition struct {
